pyproject.toml
[project]
name = "csiro-image2biomass-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core==1.3.2",
    "numpy==1.26.4",
    "pandas==2.2.3",
    "polars==1.25.0",
    "omegaconf==2.3.0",
    "kaggle==1.6.17",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "seaborn>=0.13.2",
    "joblib==1.3.2",
    "scipy==1.15.2",
    "plotly>=6.1.2",
    "scikit-learn==1.2.2",
    "wandb>=0.19.11",
    "kagglehub>=0.3.12",
    "streamlit>=1.45.1",
    "torchinfo==1.8.0",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "ruff>=0.12.5",
    "ipywidgets>=8.1.7",
    "colorama>=0.4.6",
    "palettable>=3.3.3",
    "nbformat>=5.10.4",
    "entmax>=1.3",
    "lightgbm==4.5.0",
    "xgboost==2.0.3",
    "catboost==1.2.8",
    "schedulefree>=1.4.1",
    "timm>=1.0.22",
    "albumentations>=2.0.8",
]
